update test-dds-watcher to use fork & closure
